+++ src/acpi_intel_sst.h
@@ -228,6 +228,7 @@ void	sst_dsp_set_regs_defaults(struct sst_softc *sc);
 /* acpi_intel_sst.c */
 int	sst_dsp_recover(struct sst_softc *sc);
 int	sst_dsp_core_reset(struct sst_softc *sc);
+void	sst_shim_configure(struct sst_softc *sc);
 
 /* sst_power.c */
 int	sst_reset(struct sst_softc *sc);
@@ -245,6 +246,7 @@ void	sst_dump_pch_state(struct sst_softc *sc);
 
 /* sst_sram.c */
 void	sst_sram_sanitize(struct sst_softc *sc);
+void	sst_sram_clear(struct sst_softc *sc);
 int	sst_sram_probe(struct sst_softc *sc);
 bool	sst_test_bar0(struct sst_softc *sc);
 void	sst_scan_bar0(struct sst_softc *sc);